Split test files into valid and invalid
